+++ psdraw3/Global.cpp
@@ -25,10 +25,10 @@ int cur_x=0, cur_y=0; // ubicacion del raton (durante el passive_motion_cb esta 
 bool panning=false; // indica si se esta moviendo el dibujo, para el motion
 bool selecting_zoom=false; // para hacer zoom en un area marcada, con el boton del medio, m_x0 y m_y0 guardan la primer esquina
 bool selecting_entities=false; // para selecciones múltiples (rectangulares, shift+drag derecho)
-float d_dx=0; // "paning" del dibujo
-float d_dy=0; // "paning" del dibujo
-float d_zoom=0.1; // zoom del dibujo
-float zoom=0.1; // zoom final
+double d_dx=0; // "paning" del dibujo
+double d_dy=0; // "paning" del dibujo
+double d_zoom=0.1; // zoom del dibujo
+double zoom=0.1; // zoom final
 int line_width_flechas=2, line_width_bordes=1;
 
 // auxiliares varios
